Fix picture, people page order, mostly fix layout issues
Upload landscape photo for Conner (me). Implement ordering of site.people via 'order' front matter, which is the sort front matter renamed. (It sharing the name sort with the sort function was causing some sort of infinite loop). Text layout is mostly fixed, but some text is running under my photo and isn't aligned with the rest of the text on the page